requirements.txt
streamlit>=1.28.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
googlemaps>=4.10.0
amadeus>=9.0.0

